@@ -3,14 +3,24 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@craco/craco": "^5.9.0",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
+    "@mui/material": "^6.2.1",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
-    "axios": "^0.20.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.3.4",
-    "react-scripts": "3.4.3"
+    "axios": "^1.7.9",
+    "bootstrap": "^5.3.3",
+    "markdown-to-jsx": "^7.7.1",
+    "path-browserify": "^1.0.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-markdown": "^6.0.3",
+    "react-router-dom": "^7.0.2",
+    "react-scripts": "^5.0.1",
+    "unified": "^10.1.2",
+    "vfile": "^5.3.7"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -32,5 +42,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11"
   }
 }
